+++ src/contentReader.ts
@@ -1,0 +1,4 @@
+import * as vscode from 'vscode';
+export interface ContentReader {
+    removeFromCache(uri:vscode.Uri):void;
+}